Added another Detector Constructor ( IronCylinder_geo.cpp ) taken from example and the corresponding XML file compact/IronCylinder.xml
Its a very good and simple example to understand the concept of module.
Module actually replicate the multiple copies of detector.
In the current example it replicate the DD4hep_CylinderShell, which
is actually detector constructor for Polycone, which is already built
and linked to DD4hep libraries. Its a very good way to create complex
shapes like Polycone, Polyhedra, so that we do not have to code them
in our Detector Constructor.